requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.13.4
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1739515848642/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
docker-pycreds==0.4.0
filelock @ file:///work/perseverance-python-buildout/croot/filelock_1701733993137/work
fire==0.7.0
fonttools==4.57.0
fsspec==2025.3.2
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
gmpy2 @ file:///croot/gmpy2_1738085463648/work
idna @ file:///croot/idna_1714398848350/work
imageio==2.37.0
Jinja2 @ file:///croot/jinja2_1741710844255/work
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
littleutils==0.2.4
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib==3.10.1
medmnist==3.0.1
medmnistc==0.1.0
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///work/perseverance-python-buildout/croot/mpmath_1698864994882/work
networkx @ file:///croot/networkx_1737039604450/work
numpy @ file:///croot/numpy_and_numpy_base_1742570410886/work/dist/numpy-2.2.4-cp312-cp312-linux_x86_64.whl#sha256=cd87ab48437abde52205a952c8a8e67f54f7a68985c0130688c5bdbeaf055dfc
ogb==1.3.6
opencv-python==4.11.0.86
outdated==0.2.2
packaging==24.2
pandas==2.2.3
pillow @ file:///croot/pillow_1738010226202/work
platformdirs==4.3.7
protobuf==3.20.3
psutil==7.0.0
pydantic==2.11.2
pydantic_core==2.33.1
pyparsing==3.2.3
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
requests @ file:///croot/requests_1730999120400/work
scikit-image==0.23.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentry-sdk==2.25.1
setproctitle==1.3.5
setuptools==72.1.0
six==1.17.0
smmap==5.0.2
soupsieve==2.7
sympy==1.13.1
tabulate==0.9.0
tensorboard-logger==0.1.0
termcolor==3.0.1
threadpoolctl==3.6.0
tifffile==2025.3.30
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
typing-inspection==0.4.0
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.2
urllib3 @ file:///croot/urllib3_1737133630106/work
Wand==0.6.13
wandb==0.19.9
wheel==0.45.1
wilds==2.0.0
